Get it working
Changed SPI to use 8 bits for messages which fixed the display

Add `drawRectTextured` function which takes an array of colors instead of a single solid color.

Added `setRotation` which changes the orientation the screen will render.

Added `setInverted` but it doesn't seem to work. Won't be needing it regardless so fixing it isn't high priority..

Properly exposed the huart2/hspi2/hspi3 variables as externs in `main.h`

Added my uart library (converted to cpp) for debugging. I connect to it using minicom (connect with the Justfile command).

Currently the display has an idle testing animation displaying different colors and rendering orientations. The speed isn't fantastic for rendering the whole screen (takes ~1.5s for the whole screen to draw) but it'll be fine for a few small elements needing hydration for tic tac toe.