Redo lcd rendering
Instead of having a ton of different methods that coalesce into a single lower level one, there's one method (drawRectTextured) that now is flexible in that you can pass any resolver for the color or texture mapping you want.

So this allows us to do color mixing (via a ColorResolver) or texture mapping like tiling and stretching (via a TextureResolver).

Also redid the base rendering for the game