Implementation with uart
This changes a lot of things. But the main part is that uart is used with wasd controls to select which tile to place a cross or circle on. The game loop is done pretty much. Only thing that needs to be done optionally is more user-friendly changes or something to make it clear when someone wins instead of just resetting the game loop.